+++ python/egglog/__init__.py
@@ -1,3 +1,5 @@
-from . import config  # noqa: F401
+from . import config, ipython_magic  # noqa: F401
 from .builtins import *  # noqa: F401
 from .egraph import *  # noqa: F401
+
+del ipython_magic